--- src/sunpy_sphinx_theme/sunpy/footer.html
@@ -1,42 +1,0 @@
-<footer class="footer">
-    <div class="container">
-        <div class="left-footer">
-            <p>
-                {%- if show_copyright %}
-                {%- if hasdoc('copyright') %}
-                {% trans path=pathto('copyright'), copyright=copyright|e %}&copy; <a href="{{ path }}"></a>
-                {{ copyright }}{% endtrans %}
-                {%- else %}
-                {% trans copyright=copyright|e %}&copy; {{ copyright }} {% endtrans %}
-                {%- endif %}
-                {%- endif %}
-                {{ theme_copyright_html }}
-            </p>
-        </div>
-        <div class="right-footer">
-            {%- if last_updated %}
-            {% trans last_updated=last_updated|e %}Updated on {{ last_updated }}{% endtrans %}
-            {%- if show_sphinx %}
-            {% trans sphinx_version=sphinx_version|e %}, built with <a href="http://sphinx-doc.org/">Sphinx
-                {{ sphinx_version }}</a>{% endtrans %}
-            {%- else %}
-
-            {%- endif %}
-            {%- endif %}
-            {%- if show_sphinx and not last_updated %}
-            {% trans sphinx_version=sphinx_version|e %}Built with <a href="http://sphinx-doc.org/">Sphinx
-                {{ sphinx_version }}.</a>{% endtrans %}
-            {%- endif %}
-        </div>
-        <div class="center-footer ">
-            {% if theme_footer_links %}
-            {%- for link in theme_footer_links %}
-            <a href="{{ pathto(*link[1:]) }}">{{ link[0] }}</a>
-            {% if not loop.last %}
-            <span> • </span>
-            {% endif %}
-            {%- endfor %}
-            {% endif %}
-        </div>
-    </div>
-</footer>
