--- src/sunpy_sphinx_theme/sunpy/layout.html
@@ -1,21 +1,0 @@
-{# Import the theme's layout #}
-{% extends "bootstrap/layout.html" %}
-
-{# Extra JS files #}
-{%- block scripts %}
-{{ super() }}
-<script type="text/javascript" src="{{ pathto('_static/copybutton.js', 1) }}"></script>
-<script type="text/javascript" src="{{ pathto('_static/version.js', 1) }}"></script>
-{%- endblock %}
-
-{# Custom CSS overrides #}
-{% set css_files = css_files + ['_static/sunpy_style.css'] %}
-
-{%- block extrahead -%}
-{{ super() }}
-<title> {{ docstitle|striptags }} </title>
-{%- endblock %}
-
-{%- block footer %}
-{% include "footer.html" %}
-{%- endblock %}
