--- src/sunpy_sphinx_theme/sunpy/templates/autosummary/attribute.rst
@@ -1,9 +1,0 @@
-:orphan:
-
-{{ fullname | escape | underline}}
-
-.. currentmodule:: {{ module }}
-
-attribute
-
-.. auto{{ objtype }}:: {{ objname }}
