--- src/sunpy_sphinx_theme/sunpy/templates/autosummary/member.rst
@@ -1,9 +1,0 @@
-:orphan:
-
-{{ fullname | escape | underline}}
-
-.. currentmodule:: {{ module }}
-
-member
-
-.. auto{{ objtype }}:: {{ objname }}
