--- src/sunpy_sphinx_theme/sunpy/templates/autosummary/method.rst
@@ -1,9 +1,0 @@
-:orphan:
-
-{{ fullname | escape | underline}}
-
-.. currentmodule:: {{ module }}
-
-method
-
-.. auto{{ objtype }}:: {{ objname }}
