--- src/sunpy_sphinx_theme/sunpy/templates/autosummary/minimal_module.rst
@@ -1,6 +1,0 @@
-{{ fullname | escape | underline}}
-
-.. automodule:: {{ fullname }}
-
-   {% block docstring %}
-   {% endblock %}
