--- src/sunpy_sphinx_theme/sunpy/theme.conf
@@ -1,28 +1,0 @@
-[theme]
-inherit=bootstrap
-
-[options]
-# Settings from the bootstrap theme
-navbar_pagenav = False
-source_link_position = False
-bootswatch_theme = flatly
-bootstrap_version = 3
-
-# Defaults for SunPy projects
-copyright_html =
-default_toctree_depth = 3
-github_repo = sunpy
-github_user = sunpy
-logo_url = https://sunpy.org/
-on_rtd =
-# a dict mapping path prefixes to toctree depths
-page_toctree_depths =
-social_buttons = True
-to_be_indexed = ['stable', 'latest']
-twitter_id = sunpyproject
-
-# These should be left empty and set in conf.py
-about_links =
-footer_links =
-html_sidebars =
-navbar_docs =
