+++ src/sunpy_sphinx_theme/theme/sunpy/static/copybutton.js
