+++ src/sunpy_sphinx_theme/theme/sunpy/static/img/favicon-32.ico
