+++ src/sunpy_sphinx_theme/theme/sunpy/static/img/sunpy-bg.jpg
