+++ src/sunpy_sphinx_theme/theme/sunpy/static/img/sunpy_icon_128x128.png
