+++ src/sunpy_sphinx_theme/theme/sunpy/static/sunpy_style.css
