+++ src/sunpy_sphinx_theme/theme/sunpy/static/version.js
