+++ src/sunpy_sphinx_theme/theme/sunpy/templates/navbar_start.html
@@ -1,0 +1,6 @@
+<a class="navbar-brand" href="{{ theme_logo_url }}">
+  {%- block sidebarlogo %}
+  {%- if logo %}<span><img src="{{ pathto('_static/' + logo, 1) }}"></span>{%- endif %}
+  {%- endblock %}
+  SunPy
+</a>
