+++ src/sunpy_sphinx_theme/theme/sunpy/theme.conf
@@ -1,0 +1,7 @@
+[theme]
+inherit = sphinx_book_theme
+
+[options]
+navbar_links =
+navbar_start = templates/navbar_start.html
+navbar_center = templates/navbar_center.html
